Return JSON to generate a PDF in the new API
We are extracting PDF generation out of this app into a standalone
microservice.  This application will return the JSON required to
generate this PDF.

This is currently feature flagged and can be turned on by setting a
config variable in the submitter:

PDF_SERVICE_FLAG true

If this flag is not set, it will fall back to the old PDF generating
functionality.