Add flag for including PDF and attachments in payload
This is the first small step towards promoting the submission data to the root
level of the payload.

With these flags we can figure out whether the 'user' recipient should receive
attachments (off by default), and a confirmation PDF (configured in the form)

Once this works the next step is to duplicate the payload to the root level and
use these conditionals to figure out which attachments and pdfs to send.