Attach form data to root of payload
This is a duplicate of the pdf payload for now.
Once the submitter uses this new form data, we can remove the email / pdf
specific payload.

The PDF doesn't need to be aware of the recipient.  Redacting values
isn't a feature we need on any of our current forms so remove it.

Be more explicit about the PDF payload object, instead of destructuring
it in the argument to the submitter client.