Make sure CSV submissions can accept multiple destinations
In the runner node the to field for csv attachments does not split
multiple email addresses on the comma therefore when the payload reaches
AWS SES it throws and InvalidParameter exception.

This causes the submission to get put on the failed queue which in turn
keeps trying a number of times.