Add njsscan config
This adds a github action for doing additional security scans across the
runner code base.

The initial scan was notifying of a possible injection issue. However
this is not the case as the runner is looking up JSON paths to objects
in files and is not directly connected to any SQL database.