Reduce default session timeout
This was previously increased to cater for forms that have large text
fields and may take some time for the user to input the required
information.

There are not many forms with this requirement so we can set the session
timeout individually for those particular forms without having to do it
globally.